refactor: add attraction blueprint
